requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
boto3==1.28.67
botocore==1.31.67
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.3
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==24.2
django-heroku==0.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.2
pathspec==0.10.1
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.1.9
whitenoise==6.6.0
